+++ crates/match-bindings/src/module_bindings/command_receipt_type.rs
@@ -10,7 +10,7 @@ use super::receipt_status_type::ReceiptStatus;
 #[sats(crate = __lib)]
 pub struct CommandReceipt {
     pub receipt_key: u128,
-    pub player_id: u8,
+    pub player_id: u16,
     pub client_command_id: u64,
     pub command_name: String,
     pub status: ReceiptStatus,
@@ -28,7 +28,7 @@ impl __sdk::InModule for CommandReceipt {
 /// Provides typed access to columns for query building.
 pub struct CommandReceiptCols {
     pub receipt_key: __sdk::__query_builder::Col<CommandReceipt, u128>,
-    pub player_id: __sdk::__query_builder::Col<CommandReceipt, u8>,
+    pub player_id: __sdk::__query_builder::Col<CommandReceipt, u16>,
     pub client_command_id: __sdk::__query_builder::Col<CommandReceipt, u64>,
     pub command_name: __sdk::__query_builder::Col<CommandReceipt, String>,
     pub status: __sdk::__query_builder::Col<CommandReceipt, ReceiptStatus>,
@@ -57,7 +57,7 @@ impl __sdk::__query_builder::HasCols for CommandReceipt {
 ///
 /// Provides typed access to indexed columns for query building.
 pub struct CommandReceiptIxCols {
-    pub player_id: __sdk::__query_builder::IxCol<CommandReceipt, u8>,
+    pub player_id: __sdk::__query_builder::IxCol<CommandReceipt, u16>,
     pub receipt_key: __sdk::__query_builder::IxCol<CommandReceipt, u128>,
 }
