+++ crates/match-bindings/src/module_bindings/mobilization_policy_table.rs
@@ -174,7 +174,7 @@ impl<'ctx> __sdk::WithUpdate for MobilizationPolicyTableHandle<'ctx> {
 /// but to directly chain method calls,
 /// like `ctx.db.mobilization_policy().player_id().find(...)`.
 pub struct MobilizationPolicyPlayerIdUnique<'ctx> {
-    imp: __sdk::UniqueConstraintHandle<MobilizationPolicy, u8>,
+    imp: __sdk::UniqueConstraintHandle<MobilizationPolicy, u16>,
     phantom: std::marker::PhantomData<&'ctx super::RemoteTables>,
 }
 
@@ -182,7 +182,7 @@ impl<'ctx> MobilizationPolicyTableHandle<'ctx> {
     /// Get a handle on the `player_id` unique index on the table `mobilization_policy`.
     pub fn player_id(&self) -> MobilizationPolicyPlayerIdUnique<'ctx> {
         MobilizationPolicyPlayerIdUnique {
-            imp: self.imp.get_unique_constraint::<u8>("player_id"),
+            imp: self.imp.get_unique_constraint::<u16>("player_id"),
             phantom: std::marker::PhantomData,
         }
     }
@@ -191,7 +191,7 @@ impl<'ctx> MobilizationPolicyTableHandle<'ctx> {
 impl<'ctx> MobilizationPolicyPlayerIdUnique<'ctx> {
     /// Find the subscribed row whose `player_id` column value is equal to `col_val`,
     /// if such a row is present in the client cache.
-    pub fn find(&self, col_val: &u8) -> Option<MobilizationPolicy> {
+    pub fn find(&self, col_val: &u16) -> Option<MobilizationPolicy> {
         self.imp.find(col_val)
     }
 }
@@ -199,7 +199,7 @@ impl<'ctx> MobilizationPolicyPlayerIdUnique<'ctx> {
 #[doc(hidden)]
 pub(super) fn register_table(client_cache: &mut __sdk::ClientCache<super::RemoteModule>) {
     let _table = client_cache.get_or_make_table::<MobilizationPolicy>("mobilization_policy");
-    _table.add_unique_constraint::<u8>("player_id", |row| &row.player_id);
+    _table.add_unique_constraint::<u16>("player_id", |row| &row.player_id);
 }
 
 #[doc(hidden)]
